+++ qupath-core-processing/src/main/java/qupath/opencv/dnn/DnnModel.java
@@ -2,7 +2,7 @@
  * #%L
  * This file is part of QuPath.
  * %%
- * Copyright (C) 2021 QuPath developers, The University of Edinburgh
+ * Copyright (C) 2021-2023 QuPath developers, The University of Edinburgh
  * %%
  * QuPath is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -22,132 +22,91 @@
 
 package qupath.opencv.dnn;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 import org.bytedeco.opencv.opencv_core.Mat;
-import org.slf4j.LoggerFactory;
 
 
 /**
- * Wrapper for a deep learning model in a pipeline using OpenCV.
+ * General interface for implementing a deep learning model in a pipeline using OpenCV.
+ * <p>
  * It can encapsulate a custom method needed to convert the input Mat(s) into the appropriate format, 
  * and the output back into one or more Mats.
  * <p>
- * This provides convenience methods to both convert and predict for three common scenarios:
+ * Implementations should provide convenience methods to both convert and predict for three common scenarios:
  * <ul>
  *   <li>Single input, single output; batch size 1</li>
  *   <li>Single or multiple inputs, single or multiple outputs; batch size 1</li>
  *   <li>Single input, single output; batch size &gt; 1</li>
  * </ul>
- * 
+ * <p>
+ * If only a single input and output are required, then only {@link #predict(Mat)}
+ * needs to be implemented.
+ * <p>
+ * <b>Note: </b>This was originally implemented in QuPath v0.3.0, but simplified for
+ *              QuPath v0.5.0. It no longer takes a generic parameter or requires
+ *              'blob' and 'prediction' functions to be defined.
+ *              This makes it easier to implement, and also to handle memory management.
+ *              If you want the old behavior, see {@link AbstractDnnModel}.
  * 
  * @author Pete Bankhead
- * @param <T> 
  * @see BlobFunction
  * @see PredictionFunction
- * @version 0.3.0
+ * @since 0.5.0
  */
-public interface DnnModel<T> extends AutoCloseable {
+public interface DnnModel extends AutoCloseable {
 		
 	/**
 	 * Default input layer name. This should be used when the input layer name is known or 
 	 * unimportant (e.g. the common case of a single input).
 	 */
-	public static final String DEFAULT_INPUT_NAME = "input";
+	String DEFAULT_INPUT_NAME = "input";
 	
 	/**
 	 * Default output layer name. This should be used when the output layer name is known or 
 	 * unimportant (e.g. the common case of a single output).
 	 */
-	public static final String DEFAULT_OUTPUT_NAME = "output";
+	String DEFAULT_OUTPUT_NAME = "output";
 	
 	/**
-	 * Get the function that can convert one or more OpenCV Mats into a blob supported by the prediction function 
-	 * for the first (or only) input.
-	 * @return
-	 */
-	public BlobFunction<T> getBlobFunction();
-	
-	/**
-	 * Get the function that can convert one or more OpenCV Mats into a blob supported by the prediction function for 
-	 * a specified input layer.
-	 * @param name 
-	 * @return
-	 */
-	public BlobFunction<T> getBlobFunction(String name);
-	
-	/**
-	 * Get the prediction function that can apply a prediction with one or more blobs as input.
-	 * @return
-	 */
-	public PredictionFunction<T> getPredictionFunction();
-	
-	/**
-	 * Convenience method to convert input image patches to a blobs, apply a {@link PredictionFunction} (optionally with multiple inputs/outputs),
-	 * and convert the output to a standard {@link Mat}.
-	 * <p>
-	 * Note that this only supports a batch size of 1. For larger batches or more control, {@link #getBlobFunction(String)} and 
-	 * {@link #getPredictionFunction()} should be used directly.
-	 * 
+	 * Prediction function that can take multiple inputs.
 	 * @param blobs
 	 * @return
 	 */
-	public default Map<String, Mat> convertAndPredict(Map<String, Mat> blobs) {
-		
-		var input = blobs.entrySet().stream().collect(Collectors.toMap(e -> e.getKey(), e -> getBlobFunction(e.getKey()).toBlob(e.getValue())));
-		var prediction = getPredictionFunction().predict(input);
-		
-		// TODO: Consider warning for multiple outputs (although shouldn't happen because conversion is also covered within this method)
-		var output = prediction.entrySet().stream().collect(Collectors.toMap(e -> e.getKey(), e -> getBlobFunction(e.getKey()).fromBlob(e.getValue()).get(0)));
-		
-		return output;
-	}
+	Map<String, Mat> predict(Map<String, Mat> blobs);
 
 	/**
-	 * Convenience method to convert a single image patch to a blob, apply the {@link PredictionFunction}, and convert the output to a standard {@link Mat}.
-	 * <p>
-	 * Note that this only supports a batch size of 1. For larger batches or more control, {@link #getBlobFunction(String)} and 
-	 * {@link #getPredictionFunction()} should be used directly.
-	 * 
+	 * Prediction function that takes a single input and gives a single output.
 	 * @param mat
 	 * @return
 	 */
-	public default Mat convertAndPredict(Mat mat) {
-		var blob = getBlobFunction().toBlob(mat);
-		var prediction = getPredictionFunction().predict(blob);
-		var output = getBlobFunction().fromBlob(prediction);
-		if (output.size() > 1)
-			LoggerFactory.getLogger(DnnModel.class).warn("Expected single output but got {} outputs - I will return the first only", output.size());
-		return output.get(0);
+	default Mat predict(Mat mat) {
+		return predict(Map.of(DEFAULT_INPUT_NAME, mat)).get(DEFAULT_OUTPUT_NAME);
 	}
 	
 	/**
-	 * Convenience method to convert one or more image patches to a blob, apply the {@link PredictionFunction}, and convert the output to standard Mats.
-	 * This method is intended for cases where the batch size should be greater than one; for a batch size of one, {@link #convertAndPredict(Mat)} can 
-	 * be used instead.
+	 * Prediction function that can take a batch of inputs and gives a corresponding
+	 * batch of outputs.
+	 * Each input is expected to have a single output.
 	 * @param mats
 	 * @return
 	 */
-	public default List<Mat> batchConvertAndPredict(Mat... mats) {
-		var blob = getBlobFunction().toBlob(mats);
-		var prediction = getPredictionFunction().predict(blob);
-		var output = getBlobFunction().fromBlob(prediction);
+	default List<Mat> batchPredict(List<? extends Mat> mats) {
+		List<Mat> output = new ArrayList<>();
+		for (var input : mats) {
+			output.add(predict(input));
+		}
 		return output;
 	}
 	
 	/**
 	 * Close this model if it will not be needed again.
 	 * Subclasses that require cleanup may override this.
-	 * A default, do-nothing implementation implementation is provided for backwards compatibility with v0.3.0.
-	 * 
-	 * @since v0.3.1
-	 * @implNote this was introduced to provide a mechanism to deal with a GPU memory leak when 
-	 *           using OpenCV and CUDA. New code using any DnnModel should call this method if it can 
-	 *           be known that the model will not be needed again in the future.
+	 * The default implementation does nothing.
 	 */
 	@Override
-	public default void close() throws Exception {}
+	default void close() throws Exception {}
 	
 }